refactor(api): :truck: move movies-search.api.service to the same folder as the db implementation
