+++ apps/api/src/movies/movies-search/movies-search.api.service.spec.ts
@@ -4,8 +4,9 @@ import { createMock } from '@golevelup/ts-jest'
 import { AxiosResponse } from 'axios'
 import { of } from 'rxjs'
 
-import { IMDBSearchResponse } from './interfaces/imdb-api.interface'
-import { ImdbApiHttpService } from './clients/imdb-api.service'
+import { IMDBSearchResponse } from '../interfaces/imdb-api.interface'
+import { ImdbApiHttpService } from '../clients/imdb-api.service'
+
 import { MoviesSearchApiService } from './movies-search.api.service'
 
 describe('MoviesSearchApiService', () => {
