+++ apps/api/src/movies/movies-search/movies-search.service.spec.ts
