+++ apps/api/src/movies/movies.module.ts
@@ -5,9 +5,9 @@ import { MoviesController } from './movies.controller'
 import { StreamingApiModule } from './clients/streaming-api.module'
 // IMDB API Service
 // import { MoviesSearchApiService } from './movies-search.api.service'
-import { MoviesSearchService } from './movies-search.service'
+import { MoviesSearchService } from './movies-search/movies-search.service'
 import { ImdbApiModule } from './clients/imdb-api.module'
-import { MoviesSearchDbService } from './movies-search-db/movies-search-db.service'
+import { MoviesSearchDbService } from './movies-search/movies-search.db.service'
 
 @Module({
   imports: [StreamingApiModule, ImdbApiModule],
